+++ cluster/submit-scripts/sbatch-template.sh
@@ -12,6 +12,13 @@ set -e
 # Commands #
 ############
 
+config=$(mktemp)
+cat << EOF > $config
+
+#{UAP_CONFIG}
+
+EOF
+
 array_jobs=(#{ARRAY_JOBS})
 
 #{PRE_JOB_COMMAND}